website_creating_assignment
Your webpage should include the following sections: Include a logo and navigation menu. Add a prominent section with a background image and centered text. Provide information about yourself with text and an image. Display a set of project cards with images, titles, and descriptions. Include a contact form or contact information. Add a footer with social media links and copyright information.